pyproject.toml
[project]
name = "klugscheiser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deepgram-sdk",
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
    "openai>=1.65.1",
    "pyaudio",
    "pyttsx3>=2.98",
]
